requirements.txt
# Core dependencies
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Telegram API
telethon==1.32.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# dbt for data transformation
dbt-core==1.7.3
dbt-postgres==1.7.3

# Object detection with YOLO
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.1.0

# Orchestration
dagster==1.5.8
dagster-webserver==1.5.8

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# JSON handling
jsonschema==4.20.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1 